refactor/move self-defined component from ui
